The basis for an almost gui ready chatting logic.
**How it's made:**

Talked to DeepSeek to find an optimal way to get the logic of a chatting
app capable of handling sessions and multiple users on the same network.

1- At first we got a very simple peer client up with using udp for
broadcasting our existense on the network and receiving requests and tcp
for recievingand sending the messages all in one calss called
"PeerClient".

2- At a later stage we wanted to do separation of concerns and so we
devidied this 1 file into 4 classes: "NetworkDiscovery", "PeerManager",
"TCPServer" and "PeerCore", each class has comments written to define
the job of each line in it.

3- Converted these classes from being cli driven to be almost gui ready
and right now waiting for the gui work to start and see what is missing
and not working well.